Fix the regular icosahedron generation code
Reduce GLOBE_DETAIL to 1 because of a strange bug in the icosahedron
generation code.

Fix the macros ICOSAHEDRON_FACES, ICOSAHEDRON_EDGES, and
ICOSAHEDRON_VERTICES.

Fix the indicies of the original faces.

Change the name of the local variable in the edge macro as to avoid
potential conflict with other macro variables.

Add in previous iteration vertices at the start of each subdivision
pass.