Improve globe rotation and manipulation
Limit the globe rotation along the X axis by 10 degrees in both
directions. Adjust Y rotation by previous X rotation value. These
changes will limit the control issues the globe algorithm when X
rotation is close to either +90 or -90.

Adjust the order X and Y rotations are applied to the view to change
the oscillation that occurs when rotating the globe from the X axis to
the Y axis.